+++ workspaces/copilot/plugins/copilot-backend/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-copilot-backend
 
+## 0.18.4
+
+### Patch Changes
+
+- b5e5991: Updated better-sqlite3 to v12
+
 ## 0.18.3
 
 ### Patch Changes