+++ workspaces/linkerd/plugins/linkerd-backend/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-linkerd-backend
 
+## 0.19.0
+
+### Minor Changes
+
+- da7de05: Backstage version bump to v1.50.3
+
 ## 0.18.1
 
 ### Patch Changes